Put directory structure in a code block
I noticed that the directory structure doesn't display correctly on the main repo's display of the README, since Markdown doesn't recognise tabs. This puts it in a code block so the intended formatting is preserved. An alternative could be to instead format it as a bulleted-list style, which Markdown will recognise.